+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/actions/SettingsAction.kt
@@ -1,0 +1,21 @@
+package com.github.rafaelldi.diagnosticsclientplugin.actions
+
+import com.github.rafaelldi.diagnosticsclientplugin.settings.DiagnosticsClientConfigurable
+import com.intellij.openapi.actionSystem.ActionUpdateThread
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.options.ShowSettingsUtil
+
+class SettingsAction : AnAction() {
+    override fun actionPerformed(event: AnActionEvent) {
+        val project = event.project ?: return
+        ShowSettingsUtil.getInstance().showSettingsDialog(project, DiagnosticsClientConfigurable::class.java)
+    }
+
+    override fun update(event: AnActionEvent) {
+        val project = event.project
+        event.presentation.isEnabledAndVisible = project != null
+    }
+
+    override fun getActionUpdateThread() = ActionUpdateThread.BGT
+}